MapRenderer: Fix another crashing bug.
I was really silly and tired putting the check at the end, if the iterator is still valid.
It made sense to me this way: As the check that the iterator is valid is most often true,
but the x and y check most often are not, it would be better to put the conditions first which
abort the chain of conditions fast.