Tileset: convert arrays to vectors.
This decreases memory consumption as not so many
structs are created, but only the needed amount.

The implementation is based on std::vector, so random
access is fast, but if there are gaps in the tile ids,
still memory is wasted as the vector allocates
memory for tilee ids from 0 to max tile id.