cmake file: make compilation easy with clang.
just use "cmake . -DCMAKE_CXX_COMPILER=clang" to get it
flare compiled with clang.
The c++ standard library will be linked automatically.
Instead of clang, you could have used clang++ as well, but this is
more foolproof.

Also remove a compile option, which is nice for gcc,
but has no effect in clang