Rewrite Animations system.
New classes introduced:
* AnimationSet: holds all animations which are defined in one file,
  i.e. in animations/hero.txt or animations/<enemy>.txt

* AnimationManager: manages all occurences of different animation sets.

The Entities only have a pointer to an animation set and a pointer to
an active animation. As different entities may have the same animations
with different states, the active animation of each entity is a
clone of the animations kept in the animationSet.

The AnimationSet was introduced to save some memory, (only need to have
the animations per enemy type plus one instance of animation per enemy,
so one enemy has less than the whole animationset, but only its own
animation).

The AnimationManager was introduced to reduce loading times significantly
as the animation definition files are only read once and then cached.