Refactor the animation system for avatar+enemies.
The animationset class is now responsible to load the images.

Changes in configuration files:
The images have to be specified in the animations definition file, not
at the enemy nor at the avatar.

No memory leaks found so far.

TODO: have a pool of the images (at AnimationManager or as static in
Animation set) to find doubles of images.

Known breakage/untested: the transformation of the avatar doesn't use the
animation class system.