Fix crash because of default copy constructor.
A hazard object has its own Animation object, so in the destructor of
the Hazard class the Animation will also be deleted.

If you do not implement the copy constructor explicitly, a default copy
constructor is created by the compiler. This constructor just copies over
all values. This copy constructor will finally lead to a crash if one
Hazard is copied and then deleted. On Deletion the attached Animation will
also be deleted, but the copied Hazard still thinks this Animation was
alive.

This commit
  * removes the need for a copy constructor, by passing Hazards as
    const reference.

  * introduces a copy constructor in the header, but it is not implemented
    yet. In case somebody just uses the copy constructor, the linker will
    fail, so the documentation in the Hazard class will be found.

  * removes unneeded variables 'rendered' and 'active_frame' in Hazards
    as there are now Animation pointers, which tell you if it needs
    rendering and the active frames are stored in the animation class.