Support passive powers
This supports "always-on" or passive powers for the hero.
Enemies can't use passive powers since they don't use a powers_list.
However, the powers_list has been moved to StatBlock from MenuPowers,
so enemy support should be possible in the future.

Passive powers will generally be effects with 0 duration.
To define a power as passive, use `passive=true`

Passive powers can't be dragged from the Powers menu.
They are also disabled in the Actionbar should they end up there.

This commit also makes some fixes for MenuActiveEffects.
Effects with a duration of 0 (aka infinite) are displayed now.
If an effect doesn't have an icon (icon=-1), it won't leave a gap.