Fixes #216.
When the animation definition file references the wrong path of the
enemy image (in case of type for example) the animation manager kept
increasing the required count for that spritesheet.

Now a boolean 'loaded' was introduced to the animation set class to handle
animations, which have no image (may be caused by wrong path or on
purpose).