Only execute non-trigger passives on spawn
Closes #274

The behavior of additive (aka stacking) effects has changed slightly.
Instead of adding to the magnitude of an exisiting effect, we simply
allow a new effect of the same id to be created. This lets us easily
remove passive additive effects cause by triggers such as blocking.

There was also a bug where if the hero had HP/MP bonus effects on spawn,
their HP/MP wouldn't be filled to the correct amount. A fix for this is
included.