Implement per-enemy loot tables
Use `loot=<id>,<chance>` in enemy definition files, where `id` is an item
id and `chance` is a percentage chance that it will be dropped.
An id of 0 represents a gold drop.

If multiple loot items have the same chance, it's sent to a second list of
ids which go through another RNG roll.