Fix Heal spell bug
The Heal spell wasn't working when the Physical stat was boosted.
This caused the spell to have the wrong value for max_hp, meaning HP would
be reduced to what it would be without the Physical stat boost.

There were two things wrong here:

1. When removing effects, it was possible that the next effect would be
skipped for 1 frame. This went unnoticed because it gets corrected on the
next frame.
2. Calculating primary stat bonuses needed to be moved before all other
stat calculations.