Add --game_data flag for setting a user-defined PATH_DATA
This should be handy for engine developers. Previously, there were two
methods for testing code in this repo against the contents of the
flare-game repo:

* Copy the executable to the flare-game repo directory.
* Install flare-game system-wide.

Now, those can be avoided by using something like this:

`./flare --game_data ../flare-game/`

This commit also moves commandline argument parsing outside of the
mainLoop() and into the beginning of main(). This will be necessary if
we want to define `engine_folder` via arguments in the future.