Ignore power id 0 when saving/loading power list
A power id of 0 means no power, so there is no point in adding that id
to the powers list. Also, the size of the powers_list vector itself
should be 0. Previously, it had a initial size of POWERSLOT_COUNT (10).
This meant that the `powers=` key in save files would infinitely expand
because we kept adding 10 zeros to that list every time the player saved.