Implement enemy-style loot tables for map events
The new formatting for the loot property in map events is as follows:

```
item id # or 'currency', x, y, quantity, percentage chance
```

An item id of 0 is equal to 'currency'.

This change has the added benifit of allowing items to drop in stacks of
more than one (e.g. 5 Health Potions).