Removed Byte order mark (BOM) from languages.txt.
This caused a bug where selecting a language in the options menu would instead
give you the language above the one you selected. Presumably it failed to
parse the comment on the first line properly due to the BOM character, thus
believing it is a language code, thus pushing everyone else down one place.