Adding the default mod again.
We did get 4 complaints and probably a lot of problems relating to the
default mod being treated special go unreported.
Hence let's readd the default mod to the game data.

Flare originally was trying to be simple. This may not be true any more
for some parts of the code (optimizing and error-handling contradict
being simple), but it should be simple for the users to play.