Add "Ardwen" quest to Family Crypt
The teleport in Family Crypt is now a one-way-trip. The objective is to
escape from the crypt. This is done by finding Ardwen's ring, placing it
in a slot on the coffin, and defeating the zombie Ardwen. Killing Ardwen
will lift a curse on the lever that controls the exit door.

To ensure the player gets trapped after taking the teleporter, the game
is saved when they land on the opposite side of the center bridge.

Ardwen's combat numbers are experimental. I think I've been making boss
combat numbers too small, so I've tried giving Ardwen a sizable boost.