Salted Field: less enemies, modified mine entrance area
Since this is the first real area with enemies, we shouldn't bombard
them with too many enemies.

Also, there was a problem with the crate near the mine entrance. It was
too easy to farm it by entering and exiting the mines. I've added
a small body of water that should slow the player's ability to farm it.