Update in-game version of cave tileset; Add teleporter to cave tileset
The in-game tileset was still using the old style of lighting/shadows.
The Tiled version had the new lighting (albeit missing the in-game water
tiles), so I just copied from there. The water tiles were re-rendered
with the new lighting since they were absent from the Tiled version.

The teleporter tiles were composited in GIMP with the existing
teleporter graphics and the cave floor tiles (16,17,20,21).