Add "grinding stone" sound effect
Original (CC0): http://freesound.org/people/novino/sounds/52603/

To make the sound fit better with our aesthetic, I speed it up,
increased the bass, and added a bit of reverb.

This sound will be useful for events that happen in dungeons and caves,
like moving walls and opening coffins. These are demonstrated in the
maps in this commit.