Repack minotaur animation to include cast_alt animation
Closes #491

Since packing time increases exponentially based on the number of
animations, I used a less efficient packing technique here. The end
result is that the minotaur spritesheet will use an additional 1.5 MB of
memory when loaded in game. I think this is a negligible amount in our
overall memory usage.