Fantasycore: Add wooden bridges to cave tileset
Credits: Stefan Beller, CC-BY-SA 3.0
https://opengameart.org/content/enhancement-to-the-cave-tileset

I made a few changes to the source art so that we'd only need 16 tiles
that could work anywhere. This was done by moving the shadow plane
closer to the first 16 tiles, resulting in a much tighter shadow. A few
issues are solved here:

- The original art had problems with overlapping shadows, creating dark
spots in between tiles.
- The water edge tiles needed to be used on the background layer, where
as the other tiles needed to be on an object layer. This caused more
issues with shadows. Some of it could be alleviated by including some
ground/water in the Blender file, but this would increase the tile count
and still leave some shadow issues.

The 16 tiles included in this commit are usable on both land and water.
They can all be placed on the same object layer. I recommend creating
a separate tile layer for these tiles so that objects like treasure
chests can be placed above them.