Faster turn delay for summoned zombie
The summoned zombie has a high movement speed, so its turn delay wasn't
timing out fast enough for it to follow its path. This is probably
something that can be fixed on the engine side. If it doesn't get fixed
there, the other enemies here should be reviewed for such behavior.

Version 0.90.02 -> 0.90.03