Redesign final Fort Nasu puzzle, fixes #711
This puzzle now follows the same logic of the puzzles that proceed it.
Notable changes include:

- The switches and tiles are located in the same room. Solves the
confusion with the mirrored room concept, meaning the hint could be
removed.
- There are two fewer switches, as that made sense for the layout. I was
going to try five switches (since the previous puzzle had 4), but I was
unable to arrange them nicely. So I ended up using 6.
- Since the south room is no longer part of the puzzle and would go
unused otherwise, the map layout has been tweaked so the player will
traverse through it.