New HP/MP bar graphics
This commit is to address problems with the HP bar's readablity (#815).

* The "tapered" borders have been removed. The fill graphics for the
bars are now plain rectangles.
* The fills have gotten a few tweaks to their color and brightness. The
health fill is now purely red.
* The left side of each fill has a bright 1px border. This helps the
fill stand out, even when the fill is small.
* The right side of the fill has been given a gradient to darken it.
This further emphasizes the "critical" side of the fill.

This commit requires engine 1.11.03 or greater. This is because the
fill graphics no longer have empty padding around them. Removing the
padding prevents it from being considered part of the fill graphics.