Change inappropriate date
I inferred that the Empyrean Campaign would take place in an alternate
1400s based on the weaponry. I ended up choosing '88' as it's my
birthyear, but that's quite late for the time period I was aiming for.

But more importantly, there's the dogwhistle interpretation of that
year, which was absolutely not my intention. So it's all around better
to just change it, as this date isn't critical to the story.