Re-work Torture Chambers map/quest (again)
This is the beginning of an effort to remove quests that "trap" the
player. Such quests aren't effective when the player has teleport
scrolls.

There was also a problem with the layout of the spike traps. Placing the
traps at the junctions allowed the player to bypass them completely by
moving in between the tiles. The new layout fixes this by forcing the
player to hit the spike trap if they make a wrong turn.

The spike trap base damage has been reduced, as these traps now apply
bleeding damage.

The warden's notes themselves have been re-worked as well. Instead of
a map of each section, players get some story with a description of the
"rules" that will help them navigate the traps.

The warden's notes can be kept and placed in the player's private stash,
or they can be placed on the bookstand at the end of the trap section in
order to disable the traps.