Empyrean: Ensure potion powers are on actionbar for new characters
Honestly can't remember when, but at some point item powers for carried
items stopped being added to the action bar for new characters. But it's
better this way, since there was little control over which slots got
populated. We already have the actionbar property for classes, so let's
just use that.