Empyrean: Fix floor gaps near wall tiles in grassland/cave maps
In the past, the wall tiles from these tilesets did not have the black
"mask" that covered the non-visible portion. This meant that we couldn't
draw floor tiles under them, as the floor would show on the back side.
However, this meant that there would be no floor on the *visible* side,
either.

But we have the masks now, so this commit fills in those gaps. The
automap rulesets have also been updated so that new maps will have the
floors filled correctly.